+++ old-code/projects/safe/src/lib/components/config-display-grid-fields-modal/public-api.ts
@@ -1,2 +1,0 @@
-export * from './config-display-grid-fields-modal.component';
-export * from './config-display-grid-fields-modal.module';