+++ old-code/projects/safe/src/lib/components/skeleton/skeleton-table/public-api.ts
@@ -1,2 +1,0 @@
-export * from './skeleton-table.component';
-export * from './skeleton-table.module';