+++ old-code/projects/safe/src/lib/components/ui/tagbox/public-api.ts
@@ -1,2 +1,0 @@
-export * from './tagbox.component';
-export * from './tagbox.module';