+++ old-code/projects/safe/src/lib/models/position-attribute.model.ts
@@ -1,8 +1,0 @@
-import { PositionAttributeCategory } from './position-attribute-category.model';
-
-/** Model for PositionAttribute object */
-export interface PositionAttribute {
-  value?: string;
-  category?: PositionAttributeCategory;
-  usersCount?: number;
-}