+++ old-code/projects/safe/src/lib/pipes/date/date.pipe.spec.ts
@@ -1,8 +1,0 @@
-import { SafeDatePipe } from './date.pipe';
-
-describe('SafeDatePipe', () => {
-  it('create an instance', () => {
-    const pipe = new SafeDatePipe();
-    expect(pipe).toBeTruthy();
-  });
-});