+++ old-code/projects/safe/src/lib/views/application-users/public-api.ts
@@ -1,2 +1,0 @@
-export * from './application-users.module';
-export * from './application-users.component';