+++ old-code/projects/safe/src/lib/views/public-api.ts
@@ -1,8 +1,0 @@
-// General
-export * from './profile/public-api';
-// Application
-export * from './application-distribution-lists/public-api';
-export * from './application-templates/public-api';
-export * from './application-users/public-api';
-export * from './application-notifications/public-api';
-export * from './views.module';